Added the simpleio package
The package added two simple functions for input. One for numbers and
one for strings. The package exists to abstract the Reader wrapped over
Stdin from the pupils.

The functions are:

ReadStringFromKeyboard

Which returns a string. The function reads everything up to the new
line. The returned string does not include the new line.

ReadNumberFromKeyboard

Which returns a int. The functions reads a string up to an including a
new line. The string is then converted to a number (base 10) before
being returned. If the string does not look like a number the function
asks the pupil to type the input in again.