Complete rewrite of the package.
This is a complete rewrite of the package. The package has been
rewritten to support:

Strings
Integers - both positive and negative
Floats - both positive and negative

The previous version did not support negative integers (or
positive integers if a + was supplied), or floats at all.

Internally the module is now based around a bufio.Scanner rather
that a bufio.Reader. The module has also been completely refactored
to make it easily testable. A test package is supplied.

The module is not idiomatic Go and is not intended to be. Likewise
the package is not go routine safe and is not intended to be.