Fix: Correctly handle the case where the scanner is empty
Correctly handle the case where the scanner is passes a empty
string for the functions ReadNumberFromKeyboard and
ReadDecimalFractionFromKeyboard.

If the scanner is passed an empty string i.e. no input, the
scanner.Scan() function returns false and internally sets the error to
EOF. But scanner.Err() will only return non-nil for errors that are not
EOF. This makes it impossible to distinguish between a natural EOF which
is not an error condition and being passed an empty sting which we do
want to treat as an error condition.

The work around is to keep track if scanner.Scan() every returns true.
If it does then the string in the scanner could not have been empty.
Otherwise the string in the scanner was empty. We catch this case before
returning from the functions and return zero for the value and the
same error message that is used for a badly formed number.

Also added test cases to handle an empty string and strings that contain
only white space for the functions ReadStringFromKeyboard,
ReadNumberFromKeyboard and ReadDecimalFractionFromKeyboard.