Add tests for multiple input values.
In the case of ReadStringFromKeyboard spaces are considered part
of the input string and so are valid. Everything up to the new line,
'\n' will be considered part of the input string.

In the case of both ReadNumberFromKeyboard and
ReadDecimalFractionFromKeyboard multiple numbers separated by spaces
are not valid. The functions treat this as invalid input and
return zero.

The test file now adds cases to show this behaviour.