Fix: Ensure exported functions repeatedly ask for input after a error.
Fixed all of the exported functions so that they repeatedly as the user
again for a valid input, if the user has typed in an invalid input.

Previously the functions only checked the users first input.

The new behaviour will display an error message to stdout, when there is
invalid input, and then wait for the users next input attempt.